requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-groq
faiss-cpu
python-dotenv
sentence-transformers
pypdf
streamlit